feat: [PL-31966]: tabbed interface for Authentication (#14625)
* feat: [PL-31966]: created Tabs in Configuration component

* feat: [PL-31966]: moved Authentication tab items to a new component

* feat: [PL-31966]: created a placeholder component for Allowlist

* feat: [PL-31966]: added styles for tabs

* feat: [PL-31966]: fixed imports

* feat: [PL-31966]: updated snapshot

* feat: [PL-31966]: added a string
