+++ src/modules/33-auth-settings/pages/Configuration/Allowlist.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const Allowlist: React.FC = () => {
+  return <div>Allowlist</div>
+}
+
+export default Allowlist
