+++ src/modules/33-auth-settings/pages/Configuration/Configuration.module.scss.d.ts
@@ -10,6 +10,7 @@ declare const styles: {
   readonly callout: string
   readonly collapseClassName: string
   readonly collapseHeaderClassName: string
+  readonly configurationTabs: string
   readonly height60: string
 }
 export default styles
