+++ src/modules/33-auth-settings/pages/Configuration/Configuration.tsx
@@ -6,118 +6,55 @@
  */
 
 import React from 'react'
-import { useParams } from 'react-router-dom'
-import { Callout } from '@blueprintjs/core'
+import { identity } from 'lodash-es'
+
+import { Layout, Tabs, Tab, Text } from '@harness/uicore'
+
 import { Page } from '@common/exports'
-import RBACTooltip from '@rbac/components/RBACTooltip/RBACTooltip'
-import { useGetAuthenticationSettings } from 'services/cd-ng'
 import { useStrings } from 'framework/strings'
-import type { AccountPathProps } from '@common/interfaces/RouteInterfaces'
-import AccountAndOAuth from '@auth-settings/pages/Configuration/AccountAndOAuth/AccountAndOAuth'
-import SAMLProvider from '@auth-settings/pages/Configuration/SAMLProvider/SAMLProvider'
-import RestrictEmailDomains from '@auth-settings/pages/Configuration/RestrictEmailDomains/RestrictEmailDomains'
-import { usePermission } from '@rbac/hooks/usePermission'
-import { ResourceType } from '@rbac/interfaces/ResourceType'
-import { PermissionIdentifier } from '@rbac/interfaces/PermissionIdentifier'
 import { NGBreadcrumbs } from '@common/components/NGBreadcrumbs/NGBreadcrumbs'
-import LDAPProvider from './LDAPProvider/LDAPProvider'
+import { useQueryParamsState } from '@common/hooks/useQueryParamsState'
+import Authentication from './Authentication'
+import Allowlist from './Allowlist'
 import css from './Configuration.module.scss'
 
-export interface PermissionRequest {
-  resourceScope: {
-    accountIdentifier: string
-  }
-  resource: {
-    resourceType: ResourceType
-  }
+export const VIEWS = {
+  AUTHENTICATION: 'authentication',
+  ALLOWLIST: 'allowlist'
 }
 
 const Configuration: React.FC = () => {
-  const params = useParams<AccountPathProps>()
-  const { accountId } = params
   const { getString } = useStrings()
-  const [updating, setUpdating] = React.useState(false)
-
-  const permissionRequest = {
-    resourceScope: {
-      accountIdentifier: accountId
-    },
-    resource: {
-      resourceType: ResourceType.AUTHSETTING
-    }
-  }
-
-  const [canEdit] = usePermission(
-    {
-      ...permissionRequest,
-      permissions: [PermissionIdentifier.EDIT_AUTHSETTING]
-    },
-    []
-  )
-
-  const {
-    data,
-    loading: fetchingAuthSettings,
-    error: errorWhileFetchingAuthSettings,
-    refetch: refetchAuthSettings
-  } = useGetAuthenticationSettings({
-    queryParams: {
-      accountIdentifier: accountId
-    }
+  const [view, setView] = useQueryParamsState<string>('view', '', {
+    serializer: identity,
+    deserializer: identity
   })
+  const defaultSelectedTabId = Object.values(VIEWS).includes(view) ? view : VIEWS.AUTHENTICATION
 
   return (
     <React.Fragment>
       <Page.Header breadcrumbs={<NGBreadcrumbs />} title={getString('authentication')} />
-      <Page.Body
-        loading={fetchingAuthSettings || updating}
-        loadingMessage={updating ? getString('authSettings.updating') : undefined}
-        error={
-          (errorWhileFetchingAuthSettings?.data as Error)?.message ||
-          errorWhileFetchingAuthSettings?.message ||
-          (data?.resource ? undefined : getString('somethingWentWrong'))
-        }
-        retryOnError={() => refetchAuthSettings()}
-      >
-        {data?.resource && (
-          <React.Fragment>
-            {!canEdit && (
-              <Callout icon={null} className={css.callout}>
-                <RBACTooltip
-                  permission={PermissionIdentifier.EDIT_AUTHSETTING}
-                  resourceType={permissionRequest.resource.resourceType}
-                  resourceScope={permissionRequest.resourceScope}
-                />
-              </Callout>
-            )}
-            <AccountAndOAuth
-              authSettings={data.resource}
-              refetchAuthSettings={refetchAuthSettings}
-              canEdit={canEdit}
-              setUpdating={setUpdating}
+      <Page.Body className={css.configurationTabs}>
+        <Layout.Horizontal>
+          <Tabs
+            id="configurationTabs"
+            defaultSelectedTabId={defaultSelectedTabId}
+            onChange={newTabId => {
+              setView(newTabId as string)
+            }}
+          >
+            <Tab
+              id={VIEWS.AUTHENTICATION}
+              title={<Text>{getString('authentication')}</Text>}
+              panel={<Authentication />}
             />
-            <SAMLProvider
-              authSettings={data.resource}
-              refetchAuthSettings={refetchAuthSettings}
-              permissionRequest={permissionRequest}
-              canEdit={canEdit}
-              setUpdating={setUpdating}
+            <Tab
+              id={VIEWS.ALLOWLIST}
+              title={<Text>{getString('authSettings.allowlist')}</Text>}
+              panel={<Allowlist />}
             />
-            <LDAPProvider
-              authSettings={data.resource}
-              refetchAuthSettings={refetchAuthSettings}
-              permissionRequest={permissionRequest}
-              canEdit={canEdit}
-              setUpdating={setUpdating}
-            />
-            <RestrictEmailDomains
-              whitelistedDomains={data.resource.whitelistedDomains || []}
-              refetchAuthSettings={refetchAuthSettings}
-              canEdit={canEdit}
-              setUpdating={setUpdating}
-            />
-          </React.Fragment>
-        )}
+          </Tabs>
+        </Layout.Horizontal>
       </Page.Body>
     </React.Fragment>
   )
