+++ src/modules/33-auth-settings/pages/Configuration/SAMLProvider/SAMLProvider.tsx
@@ -30,7 +30,7 @@ import { useDeleteSamlMetaData, useUpdateAuthMechanism, useGetSamlLoginTest } fr
 import { useSAMLProviderModal } from '@auth-settings/modals/SAMLProvider/useSAMLProvider'
 import RbacMenuItem from '@rbac/components/MenuItem/MenuItem'
 import { PermissionIdentifier } from '@rbac/interfaces/PermissionIdentifier'
-import type { PermissionRequest } from '@auth-settings/pages/Configuration/Configuration'
+import type { PermissionRequest } from '@auth-settings/pages/Configuration/Authentication'
 import { useFeature } from '@common/hooks/useFeatures'
 import { FeatureIdentifier } from 'framework/featureStore/FeatureIdentifier'
 import { FeatureWarningTooltip } from '@common/components/FeatureWarning/FeatureWarningWithTooltip'
