+++ src/modules/33-auth-settings/strings/strings.en.yaml
@@ -296,3 +296,4 @@ billingInfo:
     address: Address is required
     zipCode: Zip is required
     company: Company Name is required
+allowlist: Allowlist
